Consolidate conversion window tests
