+++ tests/unittests/test_utils.py
@@ -1,6 +1,5 @@
 import unittest
 from tap_google_ads.streams import generate_hash
-from tap_google_ads.streams import get_conversion_window
 from tap_google_ads.streams import get_query_date
 from tap_google_ads.streams import create_nested_resource_schema
 from tap_google_ads.sync import shuffle
@@ -355,48 +354,6 @@ class TestShuffleCustomers(unittest.TestCase):
         ]
         self.assertListEqual(expected, actual)
 
-
-class TestGetConversionWindow(unittest.TestCase):
-    def test_int_conversion_date_in_allowable_range(self):
-        actual = get_conversion_window({"conversion_window": 12})
-        expected = 12
-        self.assertEqual(expected, actual)
-
-    def test_str_conversion_date_in_allowable_range(self):
-        actual = get_conversion_window({"conversion_window": "12"})
-        expected = 12
-        self.assertEqual(expected, actual)
-
-    def test_conversion_date_outside_allowable_range(self):
-        with self.assertRaises(RuntimeError):
-            get_conversion_window({"conversion_window": 42})
-
-        with self.assertRaises(RuntimeError):
-            get_conversion_window({"conversion_window": "42"})
-
-    def test_non_int_or_str_conversion_date(self):
-        with self.assertRaises(RuntimeError):
-            get_conversion_window({"conversion_window": {"12": 12}})
-
-        with self.assertRaises(RuntimeError):
-            get_conversion_window({"conversion_window": [12]})
-
-    def test_empty_data_types_conversion_date_returns_default(self):
-        expected = 30
-
-        actual = get_conversion_window({"conversion_window": ""})
-        self.assertEqual(expected, actual)
-
-        actual = get_conversion_window({"conversion_window": {}})
-        self.assertEqual(expected, actual)
-
-        actual = get_conversion_window({"conversion_window": []})
-        self.assertEqual(expected, actual)
-
-    def test_None_conversion_date_returns_default(self):
-        actual = get_conversion_window({"conversion_window": None})
-        expected = 30
-        self.assertEqual(expected, actual)
 
 if __name__ == '__main__':
     unittest.main()
